Merge main into jramsay fork

+++ src/vs/workbench/contrib/notebook/browser/viewModel/baseCellViewModel.ts
@@ -4,9 +4,10 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { Emitter, Event } from 'vs/base/common/event';
-import { Disposable, IDisposable, IReference } from 'vs/base/common/lifecycle';
+import { Disposable, dispose, IDisposable, IReference } from 'vs/base/common/lifecycle';
 import { Mimes } from 'vs/base/common/mime';
 import { ICodeEditor } from 'vs/editor/browser/editorBrowser';
+import { ICodeEditorService } from 'vs/editor/browser/services/codeEditorService';
 import { IPosition } from 'vs/editor/common/core/position';
 import { Range } from 'vs/editor/common/core/range';
 import { Selection } from 'vs/editor/common/core/selection';
@@ -16,10 +17,13 @@ import { SearchParams } from 'vs/editor/common/model/textModelSearch';
 import { IResolvedTextEditorModel, ITextModelService } from 'vs/editor/common/services/resolverService';
 import { IConfigurationService } from 'vs/platform/configuration/common/configuration';
 import { IUndoRedoService } from 'vs/platform/undoRedo/common/undoRedo';
-import { CellEditState, CellFocusMode, CellViewModelStateChangeEvent, CursorAtBoundary, IEditableCellViewModel, INotebookCellDecorationOptions } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
+import { IWordWrapTransientState, readTransientState, writeTransientState } from 'vs/workbench/contrib/codeEditor/browser/toggleWordWrap';
+import { CellEditState, CellFocusMode, CursorAtBoundary, IEditableCellViewModel, INotebookCellDecorationOptions } from 'vs/workbench/contrib/notebook/browser/notebookBrowser';
+import { CellViewModelStateChangeEvent } from 'vs/workbench/contrib/notebook/browser/notebookViewEvents';
 import { ViewContext } from 'vs/workbench/contrib/notebook/browser/viewModel/viewContext';
 import { NotebookCellTextModel } from 'vs/workbench/contrib/notebook/common/model/notebookCellTextModel';
 import { CellKind, INotebookCellStatusBarItem, INotebookSearchOptions } from 'vs/workbench/contrib/notebook/common/notebookCommon';
+import { NotebookOptionsChangeEvent } from 'vs/workbench/contrib/notebook/common/notebookOptions';
 
 export abstract class BaseCellViewModel extends Disposable {
 
@@ -66,22 +70,6 @@ export abstract class BaseCellViewModel extends Disposable {
 
 	private _editState: CellEditState = CellEditState.Preview;
 
-	// get editState(): CellEditState {
-	// 	return this._editState;
-	// }
-
-	// set editState(newState: CellEditState) {
-	// 	if (newState === this._editState) {
-	// 		return;
-	// 	}
-
-	// 	this._editState = newState;
-	// 	this._onDidChangeState.fire({ editStateChanged: true });
-	// 	if (this._editState === CellEditState.Preview) {
-	// 		this.focusMode = CellFocusMode.Container;
-	// 	}
-	// }
-
 	private _lineNumbers: 'on' | 'off' | 'inherit' = 'inherit';
 	get lineNumbers(): 'on' | 'off' | 'inherit' {
 		return this._lineNumbers;
@@ -101,8 +89,10 @@ export abstract class BaseCellViewModel extends Disposable {
 		return this._focusMode;
 	}
 	set focusMode(newMode: CellFocusMode) {
-		this._focusMode = newMode;
-		this._onDidChangeState.fire({ focusModeChanged: true });
+		if (this._focusMode !== newMode) {
+			this._focusMode = newMode;
+			this._onDidChangeState.fire({ focusModeChanged: true });
+		}
 	}
 
 	protected _textEditor?: ICodeEditor;
@@ -111,10 +101,11 @@ export abstract class BaseCellViewModel extends Disposable {
 	}
 	private _editorListeners: IDisposable[] = [];
 	private _editorViewStates: editorCommon.ICodeEditorViewState | null = null;
+	private _editorTransientState: IWordWrapTransientState | null = null;
 	private _resolvedCellDecorations = new Map<string, INotebookCellDecorationOptions>();
 
-	private readonly _cellDecorationsChanged = this._register(new Emitter<{ added: INotebookCellDecorationOptions[], removed: INotebookCellDecorationOptions[] }>());
-	onCellDecorationsChanged: Event<{ added: INotebookCellDecorationOptions[], removed: INotebookCellDecorationOptions[] }> = this._cellDecorationsChanged.event;
+	private readonly _cellDecorationsChanged = this._register(new Emitter<{ added: INotebookCellDecorationOptions[]; removed: INotebookCellDecorationOptions[] }>());
+	onCellDecorationsChanged: Event<{ added: INotebookCellDecorationOptions[]; removed: INotebookCellDecorationOptions[] }> = this._cellDecorationsChanged.event;
 
 	private _resolvedDecorations = new Map<string, {
 		id?: string;
@@ -142,9 +133,28 @@ export abstract class BaseCellViewModel extends Disposable {
 
 	set dragging(v: boolean) {
 		this._dragging = v;
+		this._onDidChangeState.fire({ dragStateChanged: true });
 	}
 
 	protected _textModelRef: IReference<IResolvedTextEditorModel> | undefined;
+
+	private _inputCollapsed: boolean = false;
+	get isInputCollapsed(): boolean {
+		return this._inputCollapsed;
+	}
+	set isInputCollapsed(v: boolean) {
+		this._inputCollapsed = v;
+		this._onDidChangeState.fire({ inputCollapsedChanged: true });
+	}
+
+	private _outputCollapsed: boolean = false;
+	get isOutputCollapsed(): boolean {
+		return this._outputCollapsed;
+	}
+	set isOutputCollapsed(v: boolean) {
+		this._outputCollapsed = v;
+		this._onDidChangeState.fire({ outputCollapsedChanged: true });
+	}
 
 	constructor(
 		readonly viewType: string,
@@ -154,6 +164,7 @@ export abstract class BaseCellViewModel extends Disposable {
 		private readonly _configurationService: IConfigurationService,
 		private readonly _modelService: ITextModelService,
 		private readonly _undoRedoService: IUndoRedoService,
+		private readonly _codeEditorService: ICodeEditorService,
 		// private readonly _keymapService: INotebookKeymapService
 	) {
 		super();
@@ -163,8 +174,8 @@ export abstract class BaseCellViewModel extends Disposable {
 		}));
 
 		this._register(model.onDidChangeInternalMetadata(e => {
-			this._onDidChangeState.fire({ internalMetadataChanged: true, runStateChanged: e.runStateChanged });
-			if (e.runStateChanged || e.lastRunSuccessChanged) {
+			this._onDidChangeState.fire({ internalMetadataChanged: true });
+			if (e.lastRunSuccessChanged) {
 				// Statusbar visibility may change
 				this.layoutChange({});
 			}
@@ -175,10 +186,18 @@ export abstract class BaseCellViewModel extends Disposable {
 				this.lineNumbers = 'inherit';
 			}
 		}));
+
+		if (this.model.collapseState?.inputCollapsed) {
+			this._inputCollapsed = true;
+		}
+
+		if (this.model.collapseState?.outputCollapsed) {
+			this._outputCollapsed = true;
+		}
 	}
 
 
-
+	abstract updateOptions(e: NotebookOptionsChangeEvent): void;
 	abstract hasDynamicHeight(): boolean;
 	abstract getHeight(lineHeight: number): number;
 	abstract onDeselect(): void;
@@ -218,6 +237,10 @@ export abstract class BaseCellViewModel extends Disposable {
 			this._restoreViewState(this._editorViewStates);
 		}
 
+		if (this._editorTransientState) {
+			writeTransientState(editor.getModel(), this._editorTransientState, this._codeEditorService);
+		}
+
 		this._resolvedDecorations.forEach((value, key) => {
 			if (key.startsWith('_lazy_')) {
 				// lazy ones
@@ -238,6 +261,7 @@ export abstract class BaseCellViewModel extends Disposable {
 
 	detachTextEditor() {
 		this.saveViewState();
+		this.saveTransientState();
 		// decorations need to be cleared first as editors can be resued.
 		this._resolvedDecorations.forEach(value => {
 			const resolvedid = value.id;
@@ -248,7 +272,7 @@ export abstract class BaseCellViewModel extends Disposable {
 		});
 
 		this._textEditor = undefined;
-		this._editorListeners.forEach(e => e.dispose());
+		dispose(this._editorListeners);
 		this._editorListeners = [];
 		this._onDidChangeEditorAttachState.fire();
 
@@ -272,6 +296,14 @@ export abstract class BaseCellViewModel extends Disposable {
 		}
 
 		this._editorViewStates = this._textEditor.saveViewState();
+	}
+
+	private saveTransientState() {
+		if (!this._textEditor || !this._textEditor.hasModel()) {
+			return;
+		}
+
+		this._editorTransientState = readTransientState(this._textEditor.getModel(), this._codeEditorService);
 	}
 
 	saveEditorViewState() {
@@ -427,7 +459,7 @@ export abstract class BaseCellViewModel extends Disposable {
 			return 0;
 		}
 
-		const editorPadding = this._viewContext.notebookOptions.computeEditorPadding(this.internalMetadata);
+		const editorPadding = this._viewContext.notebookOptions.computeEditorPadding(this.internalMetadata, this.uri);
 		return this._textEditor.getTopForLineNumber(line) + editorPadding.top;
 	}
 
@@ -436,7 +468,7 @@ export abstract class BaseCellViewModel extends Disposable {
 			return 0;
 		}
 
-		const editorPadding = this._viewContext.notebookOptions.computeEditorPadding(this.internalMetadata);
+		const editorPadding = this._viewContext.notebookOptions.computeEditorPadding(this.internalMetadata, this.uri);
 		return this._textEditor.getTopForPosition(line, column) + editorPadding.top;
 	}
 
@@ -531,7 +563,7 @@ export abstract class BaseCellViewModel extends Disposable {
 				options.regex || false,
 				options.caseSensitive || false,
 				options.wholeWord ? options.wordSeparators || null : null,
-				false);
+				options.regex || false);
 		} else {
 			const lineCount = this.textBuffer.getLineCount();
 			const fullRange = new Range(1, 1, lineCount, this.textBuffer.getLineLength(lineCount) + 1);
@@ -542,7 +574,7 @@ export abstract class BaseCellViewModel extends Disposable {
 				return null;
 			}
 
-			cellMatches = this.textBuffer.findMatchesLineByLine(fullRange, searchData, false, 1000);
+			cellMatches = this.textBuffer.findMatchesLineByLine(fullRange, searchData, options.regex || false, 1000);
 		}
 
 		return cellMatches;
@@ -551,8 +583,14 @@ export abstract class BaseCellViewModel extends Disposable {
 	override dispose() {
 		super.dispose();
 
-		this._editorListeners.forEach(e => e.dispose());
-		this._undoRedoService.removeElements(this.uri);
+		dispose(this._editorListeners);
+
+		// Only remove the undo redo stack if we map this cell uri to itself
+		// If we are not in perCell mode, it will map to the full NotebookDocument and
+		// we don't want to remove that entire document undo / redo stack when a cell is deleted
+		if (this._undoRedoService.getUriComparisonKey(this.uri) === this.uri.toString()) {
+			this._undoRedoService.removeElements(this.uri);
+		}
 
 		if (this._textModelRef) {
 			this._textModelRef.dispose();
